Attribute Controller - userId pulled from SecurityContext

+++ src/main/java/jotbackend/controllers/AttributeController.java
@@ -1,6 +1,7 @@
 package jotbackend.controllers;
 
 import jotbackend.classes.Attribute;
+import jotbackend.classes.JotUserDetails;
 import jotbackend.repositories.AttributeRepository;
 
 import java.util.Optional;
@@ -12,6 +13,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,42 +25,76 @@ public class AttributeController {
     private AttributeRepository attributeRepository;
 
     @PostMapping(path = "/add")
-    public @ResponseBody String addNewContact (@RequestParam Integer userId,
-                        @RequestParam String title,
+    public ResponseEntity addNewContact (@RequestParam String title,
                         @RequestParam String description) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Integer userId = ((JotUserDetails)principal).getId();
+
         Attribute newAttribute = new Attribute();
         newAttribute.setUserId(userId);
         newAttribute.setTitle(title);
         newAttribute.setDescription(description);
-        attributeRepository.save(newAttribute);
-        return "Saved";
+        return new ResponseEntity(attributeRepository.save(newAttribute), HttpStatus.CREATED);
     }
 
     @GetMapping(path = "/all")
-    public @ResponseBody Page<Attribute> getAllAttributesByUserId(@RequestParam Integer userId,
-                                                              @RequestParam Integer pageNum,
-                                                              @RequestParam Integer pageSize,
-                                                              @RequestParam String sortField,
-                                                              @RequestParam String sortDirection) {
+    public ResponseEntity getAllAttributesByUserId(@RequestParam Integer pageNum,
+                                                   @RequestParam Integer pageSize,
+                                                   @RequestParam String sortField,
+                                                   @RequestParam String sortDirection) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Integer userId = ((JotUserDetails)principal).getId();
+
         Pageable pageable = PageRequest.of(pageNum, pageSize,
                 Sort.by(Sort.Direction.fromString(sortDirection), sortField));
-        return attributeRepository.findByUserId(userId, pageable);
+        return new ResponseEntity(attributeRepository.findByUserId(userId, pageable), HttpStatus.OK);
     }
 
     @GetMapping(path = "/{attributeId}")
-    public @ResponseBody
-    Optional<Attribute> getAttributeById(@PathVariable Integer attributeId) {
-        return attributeRepository.findById(attributeId);
+    public ResponseEntity getAttributeById(@PathVariable Integer attributeId) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Integer userId = ((JotUserDetails)principal).getId();
+
+        Optional<Attribute> attribute = attributeRepository.findById(attributeId);
+
+        if (!attribute.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else if (attribute.get().getUserId() != userId) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        return new ResponseEntity(attribute, HttpStatus.OK);
     }
 
     @PutMapping(path = "/update/{attributeId}")
-    public @ResponseBody
-    ResponseEntity updateAttribute(@PathVariable Integer attributeId,
-                                              @RequestParam Integer userId,
+    public ResponseEntity updateAttribute(@PathVariable Integer attributeId,
                                               @RequestParam String title,
                                               @RequestParam String description){
-        Attribute updatedAttribute = attributeRepository.findById(attributeId)
-                .orElseThrow(() -> new ResourceNotFoundException());
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Integer userId = ((JotUserDetails)principal).getId();
+
+        Optional<Attribute> attribute = attributeRepository.findById(attributeId);
+
+        if (!attribute.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else if (attribute.get().getUserId() != userId) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Attribute updatedAttribute = attribute.get();
         updatedAttribute.setUserId(userId);
         updatedAttribute.setTitle(title);
         updatedAttribute.setDescription(description);
@@ -67,8 +103,24 @@ public class AttributeController {
     }
 
     @DeleteMapping(path = "/delete/{attributeId}")
-    public @ResponseBody
-    void deleteAttributeById(@PathVariable Integer attributeId) {
+    public ResponseEntity deleteAttributeById(@PathVariable Integer attributeId) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Integer userId = ((JotUserDetails)principal).getId();
+
+        Optional<Attribute> attribute = attributeRepository.findById(attributeId);
+
+        if (!attribute.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else if (attribute.get().getUserId() != userId) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         attributeRepository.deleteById(attributeId);
+
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
 }
